fix warning
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@48829 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/nfs-utils/support/nfs/cacheio.c
@@ -238,7 +238,7 @@ cache_flush(int force)
 	    stb.st_mtime > now)
 		stb.st_mtime = time(0);
 	
-	sprintf(stime, "%ld\n", stb.st_mtime);
+	sprintf(stime, "%lld\n", (long long)stb.st_mtime);
 	for (c=0; cachelist[c]; c++) {
 		int fd;
 		sprintf(path, "/proc/net/rpc/%s/flush", cachelist[c]);
